requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-results==2.2.0
django-extra-views==0.14.0
django-probes==1.6.0
Faker==8.14.0
kombu==5.1.0
prompt-toolkit==3.0.20
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
gunicorn==20.0.4
redis==3.5.3
whitenoise==5.2.0